chore: 엔티티 변경 후 Integer 타입으로 수정 #7

+++ src/main/java/efub/eday/edayback/domain/day/title/controller/TitleController.java
@@ -18,7 +18,7 @@ public class TitleController {
 	private final TitleService titleService;
 
 	@GetMapping("/{member_id}")
-	public ResponseEntity<TitleResponseDto> getTitlePage(@PathVariable("member_id") Long memberId) {
+	public ResponseEntity<TitleResponseDto> getTitlePage(@PathVariable("member_id") Integer memberId) {
 		TitleResponseDto titleResponseDto = titleService.getTitlePage(memberId);
 		return ResponseEntity.ok(titleResponseDto);
 	}
